backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
psycopg[binary]==3.3.3
geoalchemy2==0.17.1
shapely==2.1.2
pydantic-settings==2.8.1
python-dotenv==1.1.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
reportlab==4.3.1
